fix(朗读): 数值卡片进入输入模式时把 Int 值转成 Float 再格式化
format() 的形参是 Float，这里传的是 Int 的 value，compileAppReleaseKotlin 直接
参数类型不匹配。同函数的另两处 format 调用已经带 toFloat()，只有输入模式这条漏了。

成因来自 0035d5e30（四个数值卡片统一 Tiny 样式），该次改完未编译验证。

+++ app/src/main/java/io/legado/app/ui/book/read/sheet/ReadAloudConfigContent.kt
@@ -440,7 +440,7 @@ private fun ReadAloudNumberSliderItem(
 
     LaunchedEffect(value) { sliderValue = value.toFloat() }
     LaunchedEffect(isInputMode) {
-        if (isInputMode) textFieldState.edit { replace(0, length, format(value)) }
+        if (isInputMode) textFieldState.edit { replace(0, length, format(value.toFloat())) }
     }
 
     // 拖动中让标题下的数值实时跟随滑块, 松手或收起才真正写回
